[#IOPID-276] applied review suggestions
