+++ src/model/nonce.ts
@@ -1,19 +1,15 @@
 import * as redis from "redis";
 import * as TE from "fp-ts/TaskEither";
 import { toError } from "fp-ts/lib/Either";
-import { pipe, flow } from "fp-ts/lib/function";
-import * as O from "fp-ts/lib/Option";
+import { pipe } from "fp-ts/lib/function";
 import * as E from "fp-ts/lib/Either";
-import { readableReportSimplified } from "@pagopa/ts-commons/lib/reporters";
 import {
   FUNCTION_PREFIX,
   falsyResponseToErrorAsync,
   singleStringReply
 } from "../utils/redis/client";
 import { Nonce } from "../generated/definitions/models/Nonce";
-import { LollipopSignatureInput } from "../generated/definitions/internal/LollipopSignatureInput";
 
-const nonceRegex = new RegExp('nonce="(.*?)"');
 export const NONCE_PREFIX = "NONCE-";
 export const DEFAULT_NONCE_EXPIRE_SEC = 60;
 
@@ -38,34 +34,10 @@ export const create = (nonce: Nonce) => (
     falsyResponseToErrorAsync(new Error("Error saving the key"))
   );
 
-export const getNonceFromSignatureInput = (
-  signatureInput: LollipopSignatureInput
-): E.Either<Error, Nonce> =>
-  pipe(
-    signatureInput.match(nonceRegex),
-    O.fromNullable,
-    // take out only the first group of the match
-    O.chainNullableK(matchArray => matchArray.at(1)),
-    O.fold(
-      () => E.left(new Error("Could not retrieve nonce from signature-input.")),
-      flow(
-        Nonce.decode,
-        E.mapLeft(
-          errors =>
-            new Error(
-              `Error while decoding nonce: [${readableReportSimplified(
-                errors
-              )}]`
-            )
-        )
-      )
-    )
-  );
-
 type GetDeleteNonceT = (
-  redis_client: redis.RedisClientType
-) => (nonce: Nonce) => TE.TaskEither<Error, true>;
-export const getDeleteNonce: GetDeleteNonceT = redis_client => nonce =>
+  nonce: Nonce
+) => (redis_client: redis.RedisClientType) => TE.TaskEither<Error, true>;
+export const getDeleteNonce: GetDeleteNonceT = nonce => redis_client =>
   pipe(
     TE.tryCatch(() => redis_client.del(prefixer(nonce)), E.toError),
     TE.chain(
